# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "纸喵软件系列之服务端",
        "title": "纸喵 wechat API",
        "termsOfService": "http://zhimiao.org",
        "contact": {
            "name": "API Support",
            "url": "http://tools.zhimiao.org",
            "email": "mail@xiaoliu.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/oplatform/manage/Lists": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "第三方平台管理"
                ],
                "summary": "获取平台列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "lastId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Platform"
                            }
                        }
                    }
                }
            }
        },
        "/oplatform/manage/Set": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "第三方平台管理"
                ],
                "summary": "设施平台信息",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.PlatformParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 1,\"msg\": \"操作成功\",\"data\": null}",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/CommitCode": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序控制"
                ],
                "summary": "提交代码",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.CommitCodeParam"
                        }
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/GetAuthorizerInfo": {
            "get": {
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "获取小程序授权信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/GetCodeCategory": {
            "get": {
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "获取已设置类目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/GetCodePageList": {
            "get": {
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "获取小程序线上页面列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/GetLatestAuditStatus": {
            "get": {
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "获取小程序最后一次审核状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/GetTestQrcode": {
            "get": {
                "tags": [
                    "小程序信息获取"
                ],
                "summary": "获取体验二维码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/Release": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序控制"
                ],
                "summary": "发布已审核通过的小程序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/SetDomain": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序控制"
                ],
                "summary": "刷新服务器、业务域名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/SpeedUpAudit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "小程序控制"
                ],
                "summary": "审核加急",
                "parameters": [
                    {
                        "description": "入参集合",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.SpeedUpAuditParam"
                        }
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/SubmitAudit": {
            "post": {
                "tags": [
                    "小程序控制"
                ],
                "summary": "提交审核",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.ApiResult"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/UndoCodeAudit": {
            "post": {
                "description": "撤销审核 每天1次每月10次",
                "tags": [
                    "小程序控制"
                ],
                "summary": "撤销审核",
                "parameters": [
                    {
                        "type": "string",
                        "description": "小程序APPID",
                        "name": "MiniProgramID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/auth": {
            "get": {
                "tags": [
                    "授权回调"
                ],
                "summary": "授权",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/notify": {
            "post": {
                "tags": [
                    "授权回调"
                ],
                "summary": "异步回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/redirect": {
            "get": {
                "tags": [
                    "授权回调"
                ],
                "summary": "同步回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/tpl/add": {
            "post": {
                "tags": [
                    "第三方平台模板"
                ],
                "summary": "添加草稿到模板",
                "parameters": [
                    {
                        "type": "string",
                        "description": "草稿编号",
                        "name": "draft_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/tpl/del": {
            "delete": {
                "tags": [
                    "第三方平台模板"
                ],
                "summary": "删除模板",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模板编号",
                        "name": "template_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/tpl/draft": {
            "get": {
                "tags": [
                    "第三方平台模板"
                ],
                "summary": "获取草稿列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/string"
                        }
                    }
                }
            }
        },
        "/oplatform/{PlatformID}/tpl/list": {
            "get": {
                "tags": [
                    "第三方平台模板"
                ],
                "summary": "获取模板列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/oplatform/{PlatformID}/tpl/pushToAuto": {
            "post": {
                "tags": [
                    "第三方平台模板"
                ],
                "summary": "将模板推送到所有自动升级小程序",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模板编号",
                        "name": "template_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "第三方平台APPID",
                        "name": "PlatformID",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "models.Platform": {
            "type": "object",
            "properties": {
                "authRedirectURL": {
                    "description": "用户授权成功回跳地址",
                    "type": "string"
                },
                "bizDomain": {
                    "description": "业务域名",
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "platformID": {
                    "description": "平台 appid",
                    "type": "string"
                },
                "platformKey": {
                    "description": "平台 消息解密key",
                    "type": "string"
                },
                "platformSecret": {
                    "description": "平台 appsecret",
                    "type": "string"
                },
                "platformToken": {
                    "description": "平台 token",
                    "type": "string"
                },
                "serverDomain": {
                    "description": "服务器域名",
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
        "req.CommitCodeParam": {
            "type": "object",
            "required": [
                "miniProgramID",
                "templateID"
            ],
            "properties": {
                "miniProgramID": {
                    "description": "小程序ID",
                    "type": "string"
                },
                "templateID": {
                    "description": "模板编号",
                    "type": "integer"
                }
            }
        },
...
        "resp.ApiResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}